cleaned app.js and built components for map.
