+++ src/components/Maps/FilteredList.jsx
@@ -1,0 +1,5 @@
+import  getCarparks   from "../../helper";
+import  allCarparks from '../../../carparks-all.json'
+const { Result } = allCarparks;
+const filteredList = getCarparks(Result);
+export default filteredList;
